Edit on processSuspend(). and main test #72
Fixing a bug in processSuspend() and making the main test more readable
and working on 4 processes.
BUG NOTE: there's a potiential bug in processResume() due to keeping the
process in suspend list as well as putting it in readylist. there's no
API in Queue that I can use to solve the problem. because dequeue will
remove the first process from suspendlist queue which is not necessary
to be the process that I'm resuming now.